Moved OnlineTimeStorage to api dir
