+++ src/main/java/mr/minecraft15/onlinetime/bukkit/OnlineTimeBukkitPlugin.java
@@ -25,6 +25,7 @@
 package mr.minecraft15.onlinetime.bukkit;
 
 import de.themoep.minedown.MineDown;
+import mr.minecraft15.onlinetime.api.OnlineTimeStorage;
 import mr.minecraft15.onlinetime.api.PlayerData;
 import mr.minecraft15.onlinetime.api.PluginProxy;
 import mr.minecraft15.onlinetime.api.PluginScheduler;
