+++ src/main/java/mr/minecraft15/onlinetime/bukkit/OnlineTimePlaceholderExpansion.java
@@ -25,6 +25,7 @@
 package mr.minecraft15.onlinetime.bukkit;
 
 import me.clip.placeholderapi.expansion.PlaceholderExpansion;
+import mr.minecraft15.onlinetime.api.OnlineTimeStorage;
 import mr.minecraft15.onlinetime.api.PlayerData;
 import mr.minecraft15.onlinetime.api.PluginProxy;
 import mr.minecraft15.onlinetime.common.*;
