+++ src/main/java/mr/minecraft15/onlinetime/common/FileOnlineTimeStorage.java
@@ -25,6 +25,7 @@
 package mr.minecraft15.onlinetime.common;
 
 import mr.minecraft15.onlinetime.api.FileStorageProvider;
+import mr.minecraft15.onlinetime.api.OnlineTimeStorage;
 
 import java.util.*;
